finished: update-imhex

+++ Update-ImHex.ps1
@@ -1,46 +1,94 @@
-param()
+# cSpell:ignore WerWolv APROG_DIR
+Param()
 
-$repo = "FredrikNoren/ungit"
+. "$((Join-Path -Path (Get-Item -Path $Profile).Directory -ChildPath "Utils.ps1"))";
 
-$releases = "https://api.github.com/repos/${repo}/releases"
+$script:Config = (Get-Config -Path (Join-Path -Path $PSScriptRoot -ChildPath "config.json"));
 
-$DownloadDest = "$($env:APROG_DIR)\ungit"
+$Repo = "WerWolv/ImHex"
+
+$Releases = "https://api.github.com/repos/${repo}/releases"
+
+$DownloadDest = "$($env:APROG_DIR)\ImHex"
+
+If (-not (Test-Path -LiteralPath $DownloadDest -PathType Container)) {
+  If (Test-Path -LiteralPath $DownloadDest -PathType Leaf) {
+    Throw "Destination path at $($DownloadDest) is a file not a folder."
+  }
+  $Null = New-Item -Path $DownloadDest -ItemType Directory;
+}
+
+$FoundTokens = ($script:Config.Tokens | Where-Object {
+    $Obj1 = $_;
+    Return $Null -ne ($Obj1.Addresses | Where-Object {
+        $Obj2 = $_;
+        Return $Releases -match $Obj2;
+      });
+  });
+
+Remove-Variable Config;
+$script:Bearer = $Null;
+If ($Null -ne $FoundTokens) {
+  $script:Bearer = (ConvertTo-SecureString -String $FoundTokens.Token -AsPlainText);
+}
 
 Write-Host "Determining latest release..."
-$RepoData = (Invoke-WebRequest "$($Releases)" | ConvertFrom-Json)[0].assets;
-$TagName = (Invoke-WebRequest "$$($Releases)" | ConvertFrom-Json)[0].tag_name;
+If ($Null -ne $script:Bearer) {
+  $RepoData = (Invoke-RestMethod -Uri "$($Releases)" -Authentication Bearer -Token $script:Bearer)[0].assets;
+  $TagName = (Invoke-RestMethod -Uri "$($Releases)" -Authentication Bearer -Token $script:Bearer)[0].tag_name;
+} Else {
+  $RepoData = (Invoke-RestMethod -Uri "$($Releases)" -Authentication None)[0].assets;
+  $TagName = (Invoke-RestMethod -Uri "$($Releases)" -Authentication None)[0].tag_name;
+}
 
-$FilesToDownloadApi = @()
-$FilesToDownload = @()
-$FileNames = @()
+[string[]]$FilesToDownloadApi = @()
+[string[]]$FilesToDownload = @()
+[string[]]$FileNames = @()
 
 ForEach ($Item in $RepoData) {
-  $FilesToDownloadApi += $Item.url
+  [string[]]$FilesToDownloadApi += $Item.url
 }
 
 ForEach ($Item in $filesToDownloadApi) {
-  $Data = (Invoke-WebRequest $Item | ConvertFrom-Json)
-  $FilesToDownload += $Data.browser_download_url
-  $FileNames += $Data.name
+  If ($Null -ne $script:Bearer) {
+    $Data = (Invoke-RestMethod -Uri $Item -Authentication Bearer -Token $script:Bearer)
+  } Else {
+    $Data = (Invoke-RestMethod -Uri $Item -Authentication Bearer None $script:Bearer)
+  }
+  [string[]]$FilesToDownload += $Data.browser_download_url;
+  [string[]]$FileNames += $Data.Name;
 }
+Remove-Variable Bearer;
 
 Write-Host "Latest release: $($TagName)";
 
-Write-Host "Dowloading latest release..."
+Write-Host "Downloading latest release..."
 
 $OsVersionToDownload = "";
 
 If ($IsWindows) {
-  $OsVersionToDownload = "win32";
+  # TODO: Prompt for Windows (Install), Windows-Portable, Windows-Portable-NoGPU
+  $OsVersionToDownload = "Windows\-Portable\-x86_64";
 } ElseIf ($IsLinux) {
-  $OsVersionToDownload = "linux";
+  # TODO: Prompt for Ubuntu, Fedora, AlmaLinux, or ArchLinux
+  $OsVersionToDownload = "Ubuntu";
 } ElseIf ($IsMacOS) {
-  $OsVersionToDownload = "darwin";
+  # TODO: Prompt for macOS, AppImage
+  $OsVersionToDownload = "macOS";
 }
 
-For ($Index = 0; $Index -lt $FilesToDownload.length; $Index++) {
-  $FileName = $fileNames[$Index]
-  Invoke-WebRequest $FilesToDownload[$Index] -Out (Join-Path -Path $DownloadDest -ChildPath $FileName)
+[string[]]$FilesToDownload = ($FilesToDownload | Where-Object { $_ -match $OsVersionToDownload });
+[string[]]$FileNames = ($FileNames | Where-Object { $_ -match $OsVersionToDownload });
+
+For ($Index = 0; $Index -lt $FilesToDownload.Length; $Index++) {
+  $FileName = $FileNames[$Index]
+
+  If ($Null -ne $script:Bearer) {
+    Invoke-WebRequest $FilesToDownload[$Index] -Out (Join-Path -Path $DownloadDest -ChildPath $FileName) -Headers @{ Authorization = "Bearer $($script:Bearer)" };
+  } Else {
+    Invoke-WebRequest $FilesToDownload[$Index] -Out (Join-Path -Path $DownloadDest -ChildPath $FileName);
+  }
+  Remove-Variable FoundTokens;
 }
 
 $FilesToKeep = @()
@@ -49,12 +97,12 @@ Function Test-MatchFileToKeep() {
   Param(
     # Specifies a PowerShell file object.
     [Parameter(Mandatory = $True, Position = 0, HelpMessage = "A PowerShell file object.", ParameterSetName = "FileObject")]
-    [ValidateNotNull]
+    [ValidateNotNull()]
     [PSObject]
     $File,
     # Specifies a path to a file.
     [Parameter(Mandatory = $True, Position = 0, HelpMessage = "A path to a file.", ParameterSetName = "PSPath")]
-    [ValidateNotNullOrEmpty]
+    [ValidateNotNullOrEmpty()]
     [Alias("PSPath")]
     [string]
     $Path
@@ -91,7 +139,7 @@ Function Test-MatchFilesToKeep() {
   param(
     # Specifies an array of PowerShell file object.
     [Parameter(Mandatory = $True, Position = 0, HelpMessage = "An array of PowerShell file object.")]
-    [ValidateNotNull]
+    [ValidateNotNull()]
     [PSObject[]]
     $Files
   )
@@ -99,13 +147,23 @@ Function Test-MatchFilesToKeep() {
   Return ($Files | Where-Object { Test-MatchFileToKeep -File $_ });
 }
 
-# The download destination items that are meant to be removed.
-$DownloadDesintationItems = (Get-ChildItem -Path $DownloadDest -Depth 0 -Recurse | Where-Object { Test-MatchFilesToKeep -File $_ });
+If ((Get-ChildItem -LiteralPath $DownloadDest | Where-Object { $FileNames -notcontains $_.Name }).Length -gt 0) {
+  # Write-Output (Get-ChildItem -Path $DownloadDest -Depth 0 -Recurse) | Out-Host;
+  # The download destination items that are meant to be removed.
+  $DownloadDestinationItems = (Get-ChildItem -Path $DownloadDest -Depth 0 -Recurse | Where-Object { $FileNames -notcontains $_.Name } | Where-Object { Test-MatchFilesToKeep -File $_.Name });
 
-ForEach ($Item in $DownloadDesintationItems) {
-  Remove-Item -Force -Recurse -Path $Item.FullName
+  ForEach ($Item in $DownloadDestinationItems) {
+    Remove-Item -Force -Recurse -Path $Item.FullName
+  }
+}
+
+If ($FilesToDownload.Length -gt 1) {
+  Write-Warning -Message "Downloaded more than one file. Please update the rest manually in $($DownloadDest)";
+  Exit 2;
 }
 
 Expand-Archive -Force -Path (Join-Path -Path $DownloadDest -ChildPath $FileNames[0]) -DestinationPath $DownloadDest
 
-Remove-Item -Force -Path (Join-Path -Path $DownloadDest -ChildPath $FileNames[0]);
+If (Test-Path -LiteralPath (Join-Path -Path $DownloadDest -ChildPath $FileNames[0])) {
+  Remove-Item -Force -Path (Join-Path -Path $DownloadDest -ChildPath $FileNames[0]);
+}
